+++ src/lib/components/hooks/containerEffects.js
@@ -71,7 +71,6 @@ export const useSelected = (selected, queryInput, typeaheadInput, onSelect) => {
       displayField = selected.displayField
     }
 
-    //TODO: Check if it's necessary to call this when value is undef
     if (typeof onSelect === 'function') onSelect(value, displayField)
   }, [selected, onSelect, queryInput, typeaheadInput])
 }