+++ examples/geo/App.jsx
@@ -86,7 +86,6 @@ const App = () => {
             placeholder={placeholder}
             plugins={[[recentSearchesPlugin, {ratio: 2, name: 'Recent', limit: maxItems}]]}
             styles={autocompleteStyles}
-            text="bar"
             typeahead={true}
             Cancel={CancelButton}
             Clear={ClearButton}