+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ### 1.1.2 (15 Mar 2022)
+- Prevent no items message from displaying whilst items are being fetched
+
+### 1.1.2 (15 Mar 2022)
 
 - Add note to README regarding onSelect being called when an item is no longer selected
 - Use autofocus HTML attribute instead of a useEffect hook